The Refacta-tarusr-rex

+++ Vanilla/Vanilla.Core/Opcodes/WorldOpcodes.cs
@@ -1,6 +1,6 @@
 ﻿namespace Vanilla.Core.Opcodes
 {
-    public enum WorldOpcodes : int
+    public enum WorldOpcodes : short
     {
         MSG_NULL_ACTION = 0x000,
         CMSG_BOOTME = 0x001,
